CORE: fixed some data types being wrong fixed guard making damage go negative fixed mobs not being able to guard fixed guard and parry conditions (target engaged)

+++ src/map/entities/battleentity.h
@@ -388,10 +388,10 @@ struct apAction_t
 	uint16			  animation;			// 12 bits
 	SPECEFFECT		  speceffect;			// 7 bits
     uint8             knockback;            // 3 bits
-	uint16			  param;				// 16 bits
+	uint32			  param;				// 17 bits
 	uint16			  messageID;			// 10 bits
     SUBEFFECT         additionalEffect;     // 10 bits
-    uint16            addEffectParam;       // 17 bits
+    uint32            addEffectParam;       // 17 bits
     uint16            addEffectMessage;     // 10 bits
     SUBEFFECT         spikesEffect;         // 10 bits
     uint16            spikesParam;          // 14 bits
